+++ crates/common_enums/src/enums.rs
@@ -2599,7 +2599,7 @@ pub enum DecisionEngineMerchantCategoryCode {
 }
 
 impl CardNetwork {
-    pub fn is_global_network(&self) -> bool {
+    pub fn is_signature_network(&self) -> bool {
         match self {
             Self::Interac
             | Self::Star
@@ -2850,6 +2850,7 @@ pub enum RequestIncrementalAuthorization {
     True,
     #[default]
     False,
+    Default,
 }
 
 #[derive(Clone, Copy, Eq, Hash, PartialEq, Debug, Serialize, Deserialize, strum::Display, ToSchema,)]
@@ -7653,6 +7654,10 @@ impl TransactionStatus {
             self,
             Self::ChallengeRequired | Self::ChallengeRequiredDecoupledAuthentication
         )
+    }
+
+    pub fn is_terminal_state(self) -> bool {
+        matches!(self, Self::Success | Self::Failure)
     }
 }