+++ crates/router/src/services/api.rs
@@ -1136,6 +1136,9 @@ impl Authenticate for api_models::payments::PaymentsConfirmIntentRequest {
 #[cfg(feature = "v2")]
 impl Authenticate for api_models::payments::ProxyPaymentsRequest {}
 
+#[cfg(feature = "v2")]
+impl Authenticate for api_models::payments::ExternalVaultProxyPaymentsRequest {}
+
 #[cfg(feature = "v1")]
 impl Authenticate for api_models::payments::PaymentsRequest {
     fn get_client_secret(&self) -> Option<&String> {